Moved wrangler/pages into src, extracted templates into own dir
