+++ packages/wrangler/src/pages/functions/identifiers.ts
