+++ packages/wrangler/src/pages/templates/template-plugin.ts
@@ -1,5 +1,5 @@
 import { match } from "path-to-regexp";
-import type { HTTPMethod } from "./routes";
+import type { HTTPMethod } from "../functions/routes";
 
 /* TODO: Grab these from @cloudflare/workers-types instead */
 type Params<P extends string = string> = Record<P, string | string[]>;
